+++ R/tipmap-package.R
@@ -1,0 +1,17 @@
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+#' @importFrom assertthat assert_that
+#' @importFrom assertthat is.count
+#' @importFrom assertthat is.flag
+#' @importFrom assertthat is.scalar
+#' @importFrom furrr future_map2
+#' @importFrom future plan
+#' @importFrom magrittr %>%
+#' @importFrom stats approx
+#' @importFrom stats optim
+#' @importFrom stats pbeta
+#' @importFrom stats qnorm
+## usethis namespace: end
+NULL